+++ packages/utils/src/index.ts
@@ -1,1 +1,3 @@
 export * from "./activity";
+export * from "./inventory";
+export * from "./dayjsmhq";