+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.0.7 - 31/05/2021
+- Added QRate History Title
+- Fixed Premium History Title
+## 0.0.6 - 31/05/2021
+- Sort Inventory Types by Amount
+## 0.0.5 - 31/05/2021
+- Add Inventory Options, Grouping
+## 0.0.4 - 27/05/2021
+- Fixed DayJS MHQ Plugin
+## 0.0.3 - 27/05/2021
+- Added DayJS MHQ Plugin
+- Added User Inventory Generation
 ## 0.0.2 - 19/05/2021
 - Improved Handling of Sub-Captures on User Activity Generation
 ## 0.0.1 - 19/05/2021